Splitting tests, closes #24.
